# cmd/api/docs/swagger.yaml
basePath: /v1
definitions:
  models.Chat:
    properties:
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      receiver_id:
        type: integer
      sender_id:
        type: integer
    required:
    - message
    - receiver_id
    - sender_id
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone_number:
        $ref: '#/definitions/sql.NullString'
    type: object
  sql.NullString:
    properties:
      string:
        type: string
      valid:
        description: Valid is true if String is not NULL
        type: boolean
    type: object
host: 167.99.246.163:7000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /auth/logout:
    post:
      consumes:
      - application/json
      description: User Sign out
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "401":
          description: Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "422":
          description: Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: User Sign out
      tags:
      - User Sign out with token
  /auth/signin:
    post:
      consumes:
      - application/json
      description: authenticate user
      parameters:
      - description: Email
        in: path
        name: email
        required: true
        type: string
      - description: Password
        in: path
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "422":
          description: Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Sign In user generating jwt token
      tags:
      - User Authentication
  /auth/signup:
    post:
      consumes:
      - application/json
      description: register user
      parameters:
      - description: Email
        in: path
        name: email
        required: true
        type: string
      - description: Password
        in: path
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "422":
          description: Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Sign up user generating jwt token
      tags:
      - User Registration
  /chat/message:
    post:
      consumes:
      - application/json
      description: Insert Message
      parameters:
      - description: Message
        in: path
        name: message
        required: true
        type: string
      - description: ReceiverID
        in: path
        name: receiver_id
        required: true
        type: string
      - description: SenderID
        in: path
        name: sender_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/models.Chat'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "422":
          description: Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Insert Message
      tags:
      - Insert Message
swagger: "2.0"